% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alluv_plot}
\alias{alluv_plot}
\title{Visualization}
\usage{
alluv_plot(conf_mat, path)
}
\value{
Plot
}
\description{
Creates an alluvial plot
}
\details{
A nice visualization of the confusion matrix is to create an alluvial diagram:
}
\examples{
library(alluvial)
tbc
}
