% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/funs.R
\name{data_generator}
\alias{data_generator}
\title{Dataset generator}
\usage{
data_generator(
  df,
  batch_size = 32,
  window_size_ms,
  window_stride_ms,
  shuffle = TRUE
)

data_generator(
  df,
  batch_size = 32,
  window_size_ms,
  window_stride_ms,
  shuffle = TRUE
)
}
\arguments{
\item{generator_prep}{A \code{tibble} generated from \code{\link[=prep_tbl]{prep_tbl()}}}
}
\value{
A function that defines how the TensorFlow graph should read
and pre-process data. Not compiled or run yet.

A function that defines how the TensorFlow graph should read
and pre-process data. Not compiled or run yet.
}
\description{
define a function called data_generator that will create the
generator depending on specified inputs.

define a function called data_generator that will create the
generator depending on specified inputs.
}
\examples{
library(tfdatasets)
library(tfdatasets)
}
