% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ds_train}
\alias{ds_train}
\title{Test set generator}
\usage{
ds_train(
  df,
  id_train,
  batch_size = 32,
  window_size_ms,
  window_stride_ms,
  shuffle
)
}
\arguments{
\item{generator_prep}{A \code{tibble} generated from \code{\link[=prep_tbl]{prep_tbl()}}}
}
\value{
A tf_dataset object defines how the TensorFlow graph should read
and pre-process test data. Not compiled or run yet.
}
\description{
define a function called ds_train that will create the
generator depending on specified inputs.
}
\examples{
library(tfdatasets)
}
