% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/funs.R
\name{id_sampler}
\alias{id_sampler}
\title{Train set sampler}
\usage{
id_sampler(df)

id_sampler(df)
}
\arguments{
\item{generator_prep}{A \code{tibble} generated from \code{\link[=prep_tbl]{prep_tbl()}}}
}
\value{
A vector

A vector
}
\description{
generates a sample of IDs of 70\% of the data set for
the data set split

generates a sample of IDs of 70\% of the data set for
the data set split
}
\examples{
tbc
tbc
}
