% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predicting}
\alias{predicting}
\title{Apply the trained model on validation dataset}
\usage{
predicting(df, id_train, batch_size, window_size_ms, window_stride_ms, model)
}
\value{
Confusion matrix
}
\description{
...
}
\details{
...
}
