% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/funs.R
\name{prep_tbl}
\alias{prep_tbl}
\title{Generator preparation.}
\usage{
prep_tbl(path)

prep_tbl(path)
}
\arguments{
\item{data_path}{Path to Speech Commands data dir.}
}
\value{
A \code{tibble}

A \code{tibble}
}
\description{
In this step we will list all audio .wav files into a tibble with 3 columns:
\itemize{
\item \code{fname}: the file name;
\item \code{class}: the label for each audio file;
\item \code{class_id}: a unique integer number starting from zero for each class - used to one-hot encode the classes.
This will be useful to the next step when we will create a generator using the tfdatasets package.
}

In this step we will list all audio .wav files into a tibble with 3 columns:
\itemize{
\item \code{fname}: the file name;
\item \code{class}: the label for each audio file;
\item \code{class_id}: a unique integer number starting from zero for each class - used to one-hot encode the classes.
This will be useful to the next step when we will create a generator using the tfdatasets package.
}
}
\examples{
library(tidyverse)
prep_tbl("data/speech_commands_v0.01/")
library(tidyverse)
prep_tbl("data/speech_commands_v0.01/")
}
