% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{train_model}
\alias{train_model}
\title{Define, compile, and train a Keras model.}
\usage{
train_model(df, id_train, batch_size, window_size_ms, window_stride_ms)
}
\value{
A trained Keras model.
}
\description{
Define, compile, and train a Keras model on the training
dataset.
}
\details{
The first time you run Keras in an R session,
TensorFlow usually prints verbose ouput such as
"Your CPU supports instructions that this TensorFlow binary was not compiled to use:"
and "OMP: Info #171: KMP_AFFINITY:". You can safely ignore these messages.
}
\examples{
library(keras)
library(tfdatasets)
tbc
}
